+++ src/main/kotlin/cn/spacexc/bilibilisdk/sdk/bangumi/info/timeline/Result.kt
@@ -2,7 +2,7 @@ package cn.spacexc.bilibilisdk.sdk.bangumi.info.timeline
 
 data class Result(
     val date: String,
-    val date_ts: Int,
+    val date_ts: Long,
     val day_of_week: Int,
     val episodes: List<Episode>,
     val is_today: Int
